requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comet-ml==3.40.0
configobj==5.0.8
contourpy==1.2.1
cycler==0.12.1
dulwich==0.21.7
everett==3.1.0
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightning-utilities==0.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
Pygments==2.17.2
pyparsing==3.1.2
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.45.0
simplejson==3.19.2
six==1.16.0
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.3.3
wrapt==1.16.0
wurlitzer==3.0.3
yarl==1.9.4
# DR TODO: update these to allow for the vision models I want 
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
einops==0.6.0
filelock==3.10.7
fonttools==4.39.3
ftfy==6.1.1
huggingface-hub>=0.14.1
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
nltk==3.8.1
numpy==1.21.5
opencv-python==4.7.0.72
packaging==23.0
pandas==1.4.4
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.33.3
typing_extensions==4.5.0
urllib3==1.26.15
vit-pytorch==1.2.0
wcwidth==0.2.6
zipp==3.15.0
comet-ml==3.40.0
